requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
feedparser==6.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1